# fix ndigits for bases > 62, and make it way faster