The C library function `sizeinbase` is valid only for bases <= 62.
So use Julia's implementation, which is valid for all bases,
and happens to be quite faster, e.g. by 5x or 10x
(the reason is in its faster workaround to the fact that even
for correct bases, `sizeinbase` can return a value which is "one too big".)

For an example of speed improvement,
```julia
julia> n = ZZ(2)^1234; m = ZZ(2)^51;

# master branch
julia> @btime ndigits($n, 61)
347.557 ns (3 allocations: 48 bytes)
209

julia> @btime ndigits($m, 12)
 291.242 ns (3 allocations: 48 bytes)
15

# PR
julia> @btime ndigits($n, 61)
 75.372 ns (1 allocation: 32 bytes)
209

julia> @btime ndigits($m, 12)
27.360 ns (0 allocations: 0 bytes)
15
```